Rename main.ts to sunburst.ts, refactor Datum playback
