+++ index.html
@@ -31,7 +31,7 @@
       <div id="diseaseDetails"></div>
     </div>
 
-    <script type="module" src="/src/main.ts"></script>
-    <script type="module" src="/src/viola.ts"></script>
+    <script type="module" src="/src/sunburst.ts"></script>
+    <script type="module" src="/src/audio/viola.ts"></script>
   </body>
 </html>
