+++ src/audio/datum.ts
@@ -1,6 +1,7 @@
 import { csv } from "d3";
+import { DatumNode } from "../types";
+import { TRANSITION_TIME } from "../sunburst";
 import csvUrl from "/seq.d3.csv?url";
-import { TRANSITION_TIME } from "./sunburst";
 
 const CSV = await csv(csvUrl);
 
@@ -13,10 +14,10 @@ let currentChakraGainNode = chakraContext.createGain();
  * Fades in a new chakra/category sound, fading out the current chakra sound if it exists
  * If the current sound is the same, ignores and continues looping
  */
-export async function playChakraSound(chakraName: string) {
+export async function playChakra(chakraName: string) {
   if (chakraName === currentChakra) return;
 
-  const path = (await import(`../assets/chakra/${chakraName}.mp3`)).default;
+  const path = (await import(`../../assets/chakra/${chakraName}.mp3`)).default;
   const audio = new Audio(path);
   const newGainNode = chakraContext.createGain();
   const source = chakraContext.createMediaElementSource(audio);
@@ -32,6 +33,32 @@ export async function playChakraSound(chakraName: string) {
 
   currentChakra = chakraName;
   currentChakraGainNode = newGainNode;
+}
+
+/**
+ * Plays a new disease sound, overlapping the current disease sound if it exists
+ */
+async function playDisease(diseaseName: string) {
+  const num = findDiseaseNum(diseaseName);
+
+  if (!num)
+    return console.error(`Disease row number not found: ${diseaseName}`);
+
+  new Audio(
+    (
+      await import(`../../assets/promoter/dna${num.padStart(3, "0")}.mp3`)
+    ).default
+  ).play();
+}
+
+export default function playDatum(d: DatumNode) {
+  // Chakra/category
+  if (d.depth === 1) {
+    if (d.data.chakra) playChakra(d.data.chakra);
+    else console.error(`Datum does not have a chakra: ${d}`);
+  }
+  // Disease/promoter
+  else if (d.depth === 2) playDisease(d.data.name);
 }
 
 function fade(
@@ -55,18 +82,7 @@ function boundVolume(volume: number) {
   return Math.min(Math.max(volume, 0), 1);
 }
 
-/**
- * Plays a new disease sound, overlapping the current disease sound if it exists
- */
-export async function playDiseaseSound(rowNumber: string) {
-  new Audio(
-    (
-      await import(`../assets/promoter/dna${rowNumber.padStart(3, "0")}.mp3`)
-    ).default
-  ).play();
-}
-
-export function getRowNumberForDisease(diseaseName: string) {
-  const diseaseEntry = CSV.find((entry) => entry.Disease === diseaseName);
-  return diseaseEntry ? diseaseEntry.index : null;
+function findDiseaseNum(diseaseName: string) {
+  const row = CSV.find((entry) => entry.Disease === diseaseName);
+  return row ? row.index : null;
 }
